Implement From trait to convert cpVect <–> tuple.
- (cpFloat, cpFloat) can be created from cpVect or &cpVect.
- cpVect can be created from (cpFloat, cpFloat).

Implementing From automatically gives us Into for free.

cpVect::to_tuple() is now deprecated.
Use `v.into()` or `From::from(v)` instead.

cpVect::to_tuple() never existed in a release of this crate, but the
master branch of the chipmunk crate uses it, so I deprecated it
instead of simply removing it.